requirements.txt
python-dotenv
pydantic
pydantic-settings
rich
pytest
psutil
pywin32
pywinauto
pyautogui
uiautomation
pycaw
comtypes
screen_brightness_control
Pillow
mss
opencv-python
easyocr
pytesseract
rapidfuzz
send2trash

# Future dependencies
PySide6
sounddevice
numpy
soundfile
keyboard
faster-whisper
ctranslate2
pyttsx3
requests
pyperclip
python-dateutil
plyer
win10toast
